% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc-functions.R
\name{ipcid}
\alias{ipcid}
\alias{ipcremove}
\title{Utilities for working with IPC objects}
\usage{
ipcid(id)

ipcremove(id)
}
\arguments{
\item{id}{character(1) identifying the counter. By default,
counters with the same \code{id}, including counters in
different processes, share the same state.}
}
\value{
\code{ipcid()} returns a character(1) identifier mangled to
    include the system process identifier, providing a convenient
    way of making an approximately unique or process-local counter.

\code{ipcremove()} returns (invisibly) \code{TRUE} if
    external resources were release or \code{FALSE} if not (e.g.,
    because the resources has already been released). Use
    \code{ipcremove()} to remove external state associated with
    \code{id}.
}
\description{
Utilities for working with IPC objects
}
\examples{
id <- ipcid("example-identifier")
id

ipcremove(id)

}
