% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcmutex.R
\name{lock}
\alias{lock}
\alias{unlock}
\alias{trylock}
\title{Inter-process locks}
\usage{
lock(id = .SHM_MUTEX_ID)

unlock(id = .SHM_MUTEX_ID)

trylock(id = .SHM_MUTEX_ID)
}
\description{
Inter-process locks
}
