% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc-functions.R
\name{ipcid}
\alias{ipcid}
\alias{ipcremove}
\title{Utilities for working with IPC objects}
\usage{
ipcid(id)

ipcremove(id)
}
\arguments{
\item{id}{character(1) (optional for \code{ipcid()}) identifier
string for mutex or counter.}
}
\value{
\code{ipcid()} returns a character(1) unique identifier,
    with \code{id} (if not missing) prepended.

\code{ipcremove()} returns (invisibly) \code{TRUE} if
    external resources were release or \code{FALSE} if not (e.g.,
    because the resources has already been released).
}
\description{
Use \code{ipcid()} to generate a unique mutex or
    counter identifier. A mutex or counter with the same \code{id},
    including those in different processes, share the same state.

\code{ipcremove()} removes external state associated
    with mutex or counters created with \code{id}.
}
\examples{
ipcid()
id <- ipcid("example-identifier")
id

ipcremove(id)

}
