% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipccounter.R
\name{yield}
\alias{yield}
\alias{value}
\alias{reset}
\title{Inter-process counter}
\usage{
yield(id)

value(id)

reset(id, n = 1)
}
\arguments{
\item{id}{character(1) counter or mutex identifier}

\item{n}{integer(1) value from which \code{yield()} will
increment.}
}
\value{
\code{yield()} returns an integer(1) value representing
    the next number in sequence. The first value returned is 1.

\code{value()} returns the value to be returned by the next
    call to \code{yield()}, without incrementing the counter. If
    the counter is not longer available, \code{yield()} returns
    \code{NA}.

\code{reset()} returns \code{n}, invisibly.
}
\description{
Inter-process counter
}
\examples{
id <- ipcid("cnt-example")

yield(id)
yield(id)

value(id)
yield(id)

reset(id, 10)
value(id)
yield(id)

ipcremove(id)

}
