% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcmutex.R
\name{locked}
\alias{locked}
\alias{lock}
\alias{try_lock}
\alias{unlock}
\title{Inter-process locks}
\usage{
locked(id)

lock(id)

try_lock(id)

unlock(id)
}
\arguments{
\item{id}{character(1) identifying the lock to be obtained.}
}
\value{
\code{locked()} returns TRUE when \code{mutex} is locked,
    and FALSE otherwise.

\code{lock()} creates a named lock, returning \code{TRUE}
    on success.

\code{trylock()} returns \code{TRUE} if the lock is
    obtained, \code{FALSE} otherwise.

\code{unlock()} returns \code{TRUE} on success,
    \code{FALSE} (e.g., because there is nothing to unlock)
    otherwise.
}
\description{
Inter-process locks
}
\examples{
id <- ipcid("mtx-example")

lock(id)
try_lock(id)
unlock(id)
try_lock(id)
locked(id)

ipcremove(id)

}
